Make UI significantly more Surface Duo friendly
Designed for custom resolution of 1024x681, provided via a
modified version of https://github.com/vcmi-mods/vcmi-extras
This will likely be uploaded separately at a later date.

This repositions most dialogs onto the right screen on the Surface Duo,
recenters fullscreen window headings to the left screen,
and left-aligns status bar text to ensure readability.

Note that some dialogs are too wide to fit well, so are either not
re-centered or may overflow into the other screen.
- The Marketplace dialog's center would need to be shifted.
- The Transformer window needs a new background

Untested or needs work:
- Combat with Tactics

Additionally requires a mod to change Hall background to properly
support the left-aligned grid and Combat toolbar to group all buttons.
Will either be built-in or shipped separately at a later date.